Set up core features for Shopping Management System
Added entities (Cart, CartItem, Order, OrderItem, Product, etc.) with relations.
Implemented repository, service and controller layers

+++ src/main/java/com/shopping/base/service/BaseService.java
@@ -6,6 +6,8 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.shopping.base.repository.BaseRepository;
+import com.shopping.shop.exception.ResourceNotFoundException;
+
 import jakarta.persistence.MappedSuperclass;
 
 @MappedSuperclass
@@ -14,43 +16,38 @@ public abstract class BaseService<T, ID extends Number> {
 	@Autowired
 	private BaseRepository<T, ID> baseRepository;
 	
+	private final Class<T> entityType;
+
+    protected BaseService(Class<T> entityType) {
+        this.entityType = entityType;
+    }
+	
 	public List<T> findAll() {
 		return baseRepository.findAll();
 	}
 
 	public T findById(ID id) {
-		return baseRepository.findById(id).orElse(null);
-	}
-
-	public T getOne(ID id) {
-		return baseRepository.getReferenceById(id);
+		return baseRepository.findById(id).orElseThrow(() -> 
+        new ResourceNotFoundException(getEntityName() + " not found with id: " + id));
 	}
 
 	public Optional<T> getById(ID id) {
 		return baseRepository.findById(id);
 	}
 
-	public T insert(T entity) {
-		return baseRepository.save(entity);
-	}
-
-	public T persist(T entity) {
-		return baseRepository.saveAndFlush(entity);
-
-	}
-
-	public T update(T entity) {
-		return baseRepository.save(entity);
-	}
-
 	public void deleteById(ID id) {
-		baseRepository.deleteById(id);
+		if (!baseRepository.existsById(id)) {
+            throw new ResourceNotFoundException(getEntityName() + " not found with id: " + id);
+        }
+		baseRepository.deleteById(id);	
 	}
 
-	public void deleteAll(List<ID> ids) {
-		ids.forEach(id -> {
-			baseRepository.deleteById(id);
-		});
+	public void deleteAll() {
+		baseRepository.deleteAll();
 	}
+	
+	 private String getEntityName() {
+	        return entityType.getSimpleName();
+	 }
 	
 }
